WIP: Dismiss scrollable content, Improved Example
