+++ example/lib/demo/widgets/cubic_page_view.dart
