+++ example/lib/demo/widgets/widgets.dart
@@ -1,6 +1,6 @@
 import 'package:dismissible_page/dismissible_page.dart';
-import 'package:example/models/models.dart';
-import 'package:example/pages/stories_wrapper.dart';
+import 'package:example/demo/models/models.dart';
+import 'package:example/demo/pages/stories_wrapper.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
