+++ lib/src/single_axis_dismissible_page.dart
@@ -51,12 +51,9 @@ class SingleAxisDismissiblePage extends StatefulWidget {
 }
 
 class _SingleAxisDismissiblePageState extends State<SingleAxisDismissiblePage>
-    with TickerProviderStateMixin, DismissibleDragMixin {
-  // late final AnimationController _moveController;
+    with TickerProviderStateMixin, _DismissiblePageMixin {
   late Animation<Offset> _moveAnimation;
   double _dragExtent = 0;
-
-  // bool _dragUnderway = false;
 
   @override
   void initState() {
@@ -122,10 +119,6 @@ class _SingleAxisDismissiblePageState extends State<SingleAxisDismissiblePage>
 
   DismissiblePageDismissDirection? get _dismissDirection =>
       _extentToDirection(_dragExtent);
-
-  // bool get _isActive {
-  //   return _dragUnderway || _moveController.isAnimating;
-  // }
 
   double get _overallDragAxisExtent {
     final size = context.size;
@@ -197,17 +190,19 @@ class _SingleAxisDismissiblePageState extends State<SingleAxisDismissiblePage>
     );
   }
 
+  double get _dismissThreshold =>
+      widget.dismissThresholds[_dismissDirection] ?? _kDismissThreshold;
+
   void _handleDragEnd([DragEndDetails? _]) {
     if (!_isActive || _moveController.isAnimating) return;
     _dragUnderway = false;
     if (!_moveController.isDismissed) {
-      if (_moveController.value >
-          (widget.dismissThresholds[_dismissDirection!] ??
-              _kDismissThreshold)) {
+      if (_moveController.value > _dismissThreshold) {
         widget.onDismissed.call();
       } else {
         _moveController
-          ..reverseDuration = widget.reverseDuration
+          ..reverseDuration =
+              widget.reverseDuration * (1 / _moveController.value)
           ..reverse();
         widget.onDragEnd?.call();
       }
@@ -266,7 +261,7 @@ class _SingleAxisDismissiblePageState extends State<SingleAxisDismissiblePage>
       onVerticalDragEnd: _directionIsXAxis ? null : _handleDragEnd,
       behavior: widget.behavior,
       dragStartBehavior: widget.dragStartBehavior,
-      child: DismissibleScrollNotification(
+      child: _DismissiblePageListener(
         onStart: (_) => _handleDragStart(),
         onUpdate: _handleDragUpdate,
         onEnd: _handleDragEnd,
@@ -295,105 +290,6 @@ class _SingleAxisDismissiblePageState extends State<SingleAxisDismissiblePage>
           },
           child: widget.child,
         ),
-      ),
-    );
-  }
-}
-
-mixin DismissibleDragMixin {
-  late final AnimationController _moveController;
-  bool _dragUnderway = false;
-  int _activePointerCount = 0;
-
-  bool get _isActive => _dragUnderway || _moveController.isAnimating;
-}
-
-class DismissibleScrollNotification extends StatelessWidget {
-  const DismissibleScrollNotification({
-    required this.parentState,
-    required this.onStart,
-    required this.onUpdate,
-    required this.onEnd,
-    required this.child,
-    this.onPointerDown,
-    this.onPointerUp,
-    Key? key,
-  }) : super(key: key);
-
-  final DismissibleDragMixin parentState;
-  final ValueChanged<Offset> onStart;
-  final ValueChanged<DragEndDetails> onEnd;
-  final ValueChanged<DragUpdateDetails> onUpdate;
-  final ValueChanged<PointerDownEvent>? onPointerDown;
-  final VoidCallback? onPointerUp;
-  final Widget child;
-
-  bool get _dragUnderway => parentState._dragUnderway;
-
-  void _startOrUpdateDrag(DragUpdateDetails? details) {
-    if (details == null) return;
-    if (_dragUnderway) {
-      onUpdate(details);
-    } else {
-      onStart(details.globalPosition);
-    }
-  }
-
-  void _updateDrag(DragUpdateDetails? details) {
-    if (details != null && details.primaryDelta != null) {
-      if (_dragUnderway) {
-        onUpdate(details);
-      }
-    }
-  }
-
-  bool _onScrollNotification(ScrollNotification scrollInfo) {
-    final shouldEndDrag = _dragUnderway &&
-        (scrollInfo.metrics.atEdge || !scrollInfo.metrics.outOfRange);
-
-    if (shouldEndDrag) {
-      onEnd(DragEndDetails());
-      return false;
-    }
-
-    // if (scrollInfo is OverscrollNotification) {
-    //   _startOrUpdateDrag(scrollInfo.dragDetails);
-    //   return false;
-    // }
-
-    if (scrollInfo is ScrollUpdateNotification) {
-      if (scrollInfo.metrics.outOfRange) {
-        _startOrUpdateDrag(scrollInfo.dragDetails);
-      } else {
-        _updateDrag(scrollInfo.dragDetails);
-      }
-      return false;
-    }
-
-    return false;
-  }
-
-  void _onPointerDown(PointerDownEvent event) {
-    parentState._activePointerCount++;
-    onPointerDown?.call(event);
-  }
-
-  void _onPointerUp(_) {
-    parentState._activePointerCount--;
-    if (_dragUnderway && parentState._activePointerCount == 0) {
-      onEnd(DragEndDetails());
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Listener(
-      onPointerDown: _onPointerDown,
-      onPointerCancel: _onPointerUp,
-      onPointerUp: _onPointerUp,
-      child: NotificationListener<ScrollNotification>(
-        onNotification: _onScrollNotification,
-        child: child,
       ),
     );
   }
